move into standalone test file
